+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Merged branches and closed issues
 
+- Gettable - Include quantify dependency versions in the diagnostics report (`ScheduleGettable.initialize_and_get_with_report`). (!832)
 - Qblox backend - Absolute amplitude tolerance of waveform index suppression set to `2/constants.IMMEDIATE_SZ_GAIN` which prevents uploading of all gain-zero waveforms. (!842)
 
 ## 0.17.0 (2023-10-20)